API CHANGE: adds tests for bind method
This commit adds tests for the `bind` method. Due to the way ldapjs
works, these tests cannot be sequential (the callback passed to
ldapjs.client is never invoked if so due to the Promise wrapper).
Therefore, we had to add support for connecting via Unix domain
sockets. Otherwise, we wouldn't know when the prior server instance
had freed up the test port.